requirements.txt
absl-py==0.11.0
aes==1.0.0
anyio==2.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
checkout==0.2.0
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==2.2.6
django-bootstrap3==14.2.0
django-filter==2.4.0
django-location-field==2.1.0
django-logit==1.1.3
django-qrcode==0.3
django-requestlogging==1.0.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
docopt==0.6.2
drf-tracking==1.5.0
dynaconf==3.1.4
editdistance==0.5.3
entrypoints==0.3
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gmail==0.6.3
google-auth==1.26.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
hvac==0.10.8
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-server==1.3.0
jupyterlab==3.0.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.2.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
knox==0.1.11
lmdb==1.1.1
loguru==0.5.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
mysql-connector-python==8.0.25
mysqlclient==2.0.3
Naked==0.1.31
nbclassic==0.2.6
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
path==15.1.0
pickleshare==0.7.5
Pillow==8.1.0
pipreqs==0.4.10
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.14.0
publicip==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pycryptodome-test-vectors==1.0.2
pycryptodomex==3.10.1
Pygments==2.8.0
PyJWT==2.0.1
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
Pytess==1.0.0
pytesseract==0.3.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2019.3
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-learn==0.24.1
scipy==1.6.0
Send2Trash==1.5.0
serializers==0.2.4
shellescape==3.8.1
shortuuid==1.0.1
six==1.12.0
sklearn==0.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
validators==0.18.2
virtualenv==20.4.2
voila==0.2.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
win32-setctime==1.0.3
wrapt==1.12.1
yarg==0.1.9
zipp==3.4.0
